CMake: add out-of-source project files
CMake follows the clean source strategy when building out-of-source.
This is violated by Qt Creator by putting a .user file into the sources
tree.

With this patch the .user and a .qtcreator file are created in the
build dir which makes it also possible to open the project with
a double click.

Change-Id: I323edec917840c65664f1a8900ab678c593be4ce